requirements.txt
requirementsFlask==0.10.1
Flask-Cors==3.0.7
Flask-SQLAlchemy==1.0
mysqlclient==1.4.2.post1
SQLAlchemy==1.3.1
SQLAlchemy-Utils==0.29.8
